requirements.txt
fastapi==0.104.1
keras==2.14.0
Pillow==10.1.0
tensorflow==2.14.0
uvicorn==0.22.0